fixed ticket_watcher.user_id datatype, moved from BIGINT to INTEGER

+++ scripts/database/otrs-schema.db2.sql
@@ -693,7 +693,7 @@ CREATE INDEX index_time_accou77 ON time_accounting (ticket_id);
 -- ----------------------------------------------------------
 CREATE TABLE ticket_watcher (
     ticket_id BIGINT NOT NULL,
-    user_id BIGINT NOT NULL,
+    user_id INTEGER NOT NULL,
     create_time TIMESTAMP NOT NULL,
     create_by INTEGER NOT NULL,
     change_time TIMESTAMP NOT NULL,
